fix: moves base nodes to src folder
also improves init
