+++ node_wrappers/mediapipe_vision/face/detection/face_detection_visualization.py
@@ -3,7 +3,7 @@ import logging
 
 import cv2
 
-from ...common.base_visualization_nodes import BaseDetectionVisualizationNode
+from .....src.mediapipe_vision.common.base_visualization_nodes import BaseDetectionVisualizationNode
 
 logger = logging.getLogger(__name__)
