+++ node_wrappers/mediapipe_vision/face/landmark/face_landmark_detection.py
@@ -5,10 +5,10 @@ import logging
 import torch
 
 from .....src.mediapipe_vision.face_landmark.detector import FaceLandmarkDetector
-from ...common.base_detector_node import BaseMediaPipeDetectorNode
+from .....src.mediapipe_vision.common.base_detector_node import BaseMediaPipeDetectorNode
 
 # Import Base Classes
-from ...common.model_loader import MediaPipeModelLoaderBaseNode
+from .....src.mediapipe_vision.common.model_loader import MediaPipeModelLoaderBaseNode
 
 logger = logging.getLogger(__name__)
