+++ node_wrappers/mediapipe_vision/hand/gesture/gesture_visualiztion.py
@@ -3,7 +3,7 @@ import logging
 
 import cv2
 
-from ...common.base_visualization_nodes import BaseDetectionVisualizationNode
+from .....src.mediapipe_vision.common.base_visualization_nodes import BaseDetectionVisualizationNode
 
 logger = logging.getLogger(__name__)
