+++ node_wrappers/mediapipe_vision/hand/gesture/guesture.py
@@ -3,10 +3,10 @@ import logging
 import torch
 
 from .....src.mediapipe_vision.gesture_recognition.detector import GestureRecognizer
-from ...common.base_detector_node import BaseMediaPipeDetectorNode
+from .....src.mediapipe_vision.common.base_detector_node import BaseMediaPipeDetectorNode
 
 # Import Base Loader and Detector
-from ...common.model_loader import MediaPipeModelLoaderBaseNode
+from .....src.mediapipe_vision.common.model_loader import MediaPipeModelLoaderBaseNode
 
 logger = logging.getLogger(__name__)
 _category = "Realtime Nodes/MediaPipe Vision/Hand/GestureRecognition"
