+++ node_wrappers/mediapipe_vision/holistic/holistic_landmark.py
@@ -3,7 +3,7 @@ import logging
 import torch
 
 from ....src.mediapipe_vision.holistic_landmark.detector import HolisticLandmarkDetector
-from ..common.base_detector_node import BaseMediaPipeDetectorNode
+from ....src.mediapipe_vision.common.base_detector_node import BaseMediaPipeDetectorNode
 
 logger = logging.getLogger(__name__)
 _category = "Realtime Nodes/MediaPipe Vision/Holistic"
