+++ node_wrappers/mediapipe_vision/image_embedding/image_embeddings.py
@@ -3,10 +3,10 @@ import logging
 import torch
 
 from ....src.mediapipe_vision.image_embedding.detector import ImageEmbedder
-from ..common.base_detector_node import BaseMediaPipeDetectorNode
+from ....src.mediapipe_vision.common.base_detector_node import BaseMediaPipeDetectorNode
 
 # Import Base Loader and Detector
-from ..common.model_loader import MediaPipeModelLoaderBaseNode
+from ....src.mediapipe_vision.common.model_loader import MediaPipeModelLoaderBaseNode
 
 logger = logging.getLogger(__name__)
 _category = "Realtime Nodes/MediaPipe Vision/ImageEmbedding"
