+++ node_wrappers/mediapipe_vision/interactive_segmentation/interactive_segmentation.py
@@ -9,7 +9,7 @@ from ....src.mediapipe_vision.interactive_segmentation.detector import (
 from ....src.mediapipe_vision.types import PointOfInterest
 
 # Import Base Loader and Detector
-from ..common.model_loader import MediaPipeModelLoaderBaseNode
+from ....src.mediapipe_vision.common.model_loader import MediaPipeModelLoaderBaseNode
 
 logger = logging.getLogger(__name__)
 _category = "Realtime Nodes/MediaPipe Vision/InteractiveSegmentation"
