+++ node_wrappers/mediapipe_vision/object_detection/object_detection.py
@@ -3,8 +3,8 @@ import logging
 import torch
 
 from ....src.mediapipe_vision.object_detection.detector import ObjectDetector
-from ..common.base_detector_node import BaseMediaPipeDetectorNode
-from ..common.model_loader import MediaPipeModelLoaderBaseNode
+from ....src.mediapipe_vision.common.base_detector_node import BaseMediaPipeDetectorNode
+from ....src.mediapipe_vision.common.model_loader import MediaPipeModelLoaderBaseNode
 
 logger = logging.getLogger(__name__)
 _category = "Realtime Nodes/MediaPipe Vision/ObjectDetection"
