+++ src/mediapipe_vision/common/base_detector_node.py
